backend/local/requirements.txt
psycopg2==2.8.3
django-simple-history==2.7.0
djangorestframework-jwt==1.11.0